✨ Set ingress host using laravel app url automatically
This is required in order to avoid repetitiveness in users' values override, and less maintenance when adding/removing ingresses objects.

This is not a breaking change; it's the opposite, it will allow a seamless upgrade with the new working ingresses without needing to update values overrides.

This will also allow to remove ingress options at all for the parent chart in our values overrides.